refactor: split matrix.client into client_session and client_room_admin (ARCH-B-3 round 2 final)
