+++ src/mindroom/matrix/cache/thread_cache_helpers.py
@@ -22,6 +22,31 @@ def latest_visible_thread_event_id(history: Sequence[ResolvedVisibleMessage]) ->
     return history[-1].visible_event_id or history[-1].event_id or None
 
 
+def thread_cache_rejection_reason(
+    cache_state: ThreadCacheState | None,
+    *,
+    runtime_started_at: float | None,
+    now: float | None = None,
+) -> str | None:
+    """Return why one durable thread snapshot must be rejected, if at all."""
+    rejection_reason: str | None = None
+    if cache_state is None:
+        rejection_reason = "no_cache_state"
+    elif cache_state.validated_at is None:
+        rejection_reason = "cache_never_validated"
+    elif runtime_started_at is not None and cache_state.validated_at < runtime_started_at:
+        rejection_reason = "validated_before_runtime_start"
+    elif cache_state.invalidated_at is not None and cache_state.invalidated_at >= cache_state.validated_at:
+        rejection_reason = "thread_invalidated_after_validation"
+    elif cache_state.room_invalidated_at is not None and cache_state.room_invalidated_at >= cache_state.validated_at:
+        rejection_reason = "room_invalidated_after_validation"
+    else:
+        checked_at = time.time() if now is None else now
+        if checked_at - cache_state.validated_at > THREAD_CACHE_MAX_AGE_SECONDS:
+            rejection_reason = "cache_too_old"
+    return rejection_reason
+
+
 def thread_cache_state_is_usable(
     cache_state: ThreadCacheState | None,
     *,
@@ -29,13 +54,11 @@ def thread_cache_state_is_usable(
     now: float | None = None,
 ) -> bool:
     """Return whether one durable thread snapshot is safe to reuse."""
-    if cache_state is None or cache_state.validated_at is None:
-        return False
-    if runtime_started_at is not None and cache_state.validated_at < runtime_started_at:
-        return False
-    if cache_state.invalidated_at is not None and cache_state.invalidated_at >= cache_state.validated_at:
-        return False
-    if cache_state.room_invalidated_at is not None and cache_state.room_invalidated_at >= cache_state.validated_at:
-        return False
-    checked_at = time.time() if now is None else now
-    return checked_at - cache_state.validated_at <= THREAD_CACHE_MAX_AGE_SECONDS
+    return (
+        thread_cache_rejection_reason(
+            cache_state,
+            runtime_started_at=runtime_started_at,
+            now=now,
+        )
+        is None
+    )
