+++ src/mindroom/matrix/client_session.py
@@ -1,0 +1,163 @@
+"""Matrix session lifecycle helpers."""
+
+from __future__ import annotations
+
+import ssl as ssl_module
+from contextlib import asynccontextmanager
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import nio
+
+from mindroom.constants import RuntimePaths, encryption_keys_dir, runtime_matrix_ssl_verify
+from mindroom.logging_config import get_logger
+
+if TYPE_CHECKING:
+    from collections.abc import AsyncGenerator
+
+logger = get_logger(__name__)
+
+_PERMANENT_MATRIX_STARTUP_ERROR_CODES = frozenset(
+    {
+        "M_FORBIDDEN",
+        "M_USER_DEACTIVATED",
+        "M_UNKNOWN_TOKEN",
+        "M_INVALID_USERNAME",
+    },
+)
+
+
+class PermanentMatrixStartupError(ValueError):
+    """Raised for Matrix startup failures that should not be retried."""
+
+
+def _require_runtime_paths_arg(runtime_paths: object) -> RuntimePaths:
+    """Reject stale positional call shapes with a clear error."""
+    if isinstance(runtime_paths, RuntimePaths):
+        return runtime_paths
+    msg = (
+        "matrix_client() requires RuntimePaths as its second argument. "
+        "Call matrix_client(homeserver, runtime_paths, user_id=...)"
+    )
+    raise TypeError(msg)
+
+
+def matrix_startup_error(
+    message: str,
+    *,
+    response: object | None = None,
+    permanent: bool = False,
+) -> ValueError:
+    """Return the appropriate startup exception type for a Matrix failure."""
+    if permanent:
+        return PermanentMatrixStartupError(message)
+    if isinstance(response, nio.ErrorResponse) and response.status_code in _PERMANENT_MATRIX_STARTUP_ERROR_CODES:
+        return PermanentMatrixStartupError(message)
+    return ValueError(message)
+
+
+def _maybe_ssl_context(homeserver: str, runtime_paths: RuntimePaths) -> ssl_module.SSLContext | None:
+    if homeserver.startswith("https://"):
+        if not runtime_matrix_ssl_verify(runtime_paths=runtime_paths):
+            ssl_context = ssl_module.create_default_context()
+            ssl_context.check_hostname = False
+            ssl_context.verify_mode = ssl_module.CERT_NONE
+        else:
+            ssl_context = ssl_module.create_default_context()
+        return ssl_context
+    return None
+
+
+def _create_matrix_client(
+    homeserver: str,
+    runtime_paths: RuntimePaths,
+    user_id: str | None = None,
+    access_token: str | None = None,
+    store_path: str | None = None,
+) -> nio.AsyncClient:
+    """Create a Matrix client with consistent configuration."""
+    runtime_paths = _require_runtime_paths_arg(runtime_paths)
+    ssl_context = _maybe_ssl_context(homeserver, runtime_paths=runtime_paths)
+
+    if store_path is None and user_id:
+        safe_user_id = user_id.replace(":", "_").replace("@", "")
+        store_path = str(encryption_keys_dir(runtime_paths=runtime_paths) / safe_user_id)
+        Path(store_path).mkdir(parents=True, exist_ok=True)
+
+    client = nio.AsyncClient(
+        homeserver,
+        user_id or "",
+        store_path=store_path,
+        ssl=ssl_context,  # ty: ignore[invalid-argument-type]
+    )
+    if user_id:
+        client.user_id = user_id
+    if access_token:
+        client.access_token = access_token
+    return client
+
+
+@asynccontextmanager
+async def matrix_client(
+    homeserver: str,
+    runtime_paths: RuntimePaths,
+    user_id: str | None = None,
+    access_token: str | None = None,
+) -> AsyncGenerator[nio.AsyncClient, None]:
+    """Context manager for Matrix client that ensures proper cleanup."""
+    runtime_paths = _require_runtime_paths_arg(runtime_paths)
+    client = _create_matrix_client(homeserver, runtime_paths, user_id, access_token)
+    try:
+        yield client
+    finally:
+        await client.close()
+
+
+async def login(
+    homeserver: str,
+    user_id: str,
+    password: str,
+    runtime_paths: RuntimePaths,
+) -> nio.AsyncClient:
+    """Login to Matrix and return an authenticated client."""
+    runtime_paths = _require_runtime_paths_arg(runtime_paths)
+    client = _create_matrix_client(homeserver, runtime_paths, user_id)
+
+    response = await client.login(password)
+    if isinstance(response, nio.LoginResponse):
+        logger.info("matrix_login_succeeded", user_id=user_id)
+        return client
+    await client.close()
+    msg = f"Failed to login {user_id}: {response}"
+    raise matrix_startup_error(msg, response=response)
+
+
+async def restore_login(
+    homeserver: str,
+    user_id: str,
+    device_id: str,
+    access_token: str,
+    runtime_paths: RuntimePaths,
+) -> nio.AsyncClient:
+    """Restore one authenticated Matrix session without creating a new device."""
+    runtime_paths = _require_runtime_paths_arg(runtime_paths)
+    client = _create_matrix_client(homeserver, runtime_paths, user_id, access_token)
+    client.restore_login(user_id, device_id, access_token)
+
+    response = await client.whoami()
+    if isinstance(response, nio.WhoamiResponse):
+        logger.info("matrix_login_restored", user_id=user_id, device_id=device_id)
+        return client
+
+    await client.close()
+    msg = f"Failed to restore Matrix login for {user_id}: {response}"
+    raise matrix_startup_error(msg, response=response)
+
+
+__all__ = [
+    "PermanentMatrixStartupError",
+    "login",
+    "matrix_client",
+    "matrix_startup_error",
+    "restore_login",
+]
