+++ tach.toml
@@ -117,10 +117,9 @@ depends_on = []
 [[modules]]
 path = "mindroom.matrix.client"
 depends_on = [
-    "mindroom.constants",
-    "mindroom.matrix.cache",
-    "mindroom.matrix.cache.thread_history_result",
     "mindroom.matrix.client_delivery",
+    "mindroom.matrix.client_room_admin",
+    "mindroom.matrix.client_session",
     "mindroom.matrix.client_thread_history",
     "mindroom.matrix.client_visible_messages",
 ]
@@ -128,6 +127,14 @@ depends_on = [
 [[modules]]
 path = "mindroom.matrix.client_delivery"
 depends_on = ["mindroom.matrix.conversation_cache"]
+
+[[modules]]
+path = "mindroom.matrix.client_room_admin"
+depends_on = []
+
+[[modules]]
+path = "mindroom.matrix.client_session"
+depends_on = ["mindroom.constants"]
 
 [[modules]]
 path = "mindroom.matrix.client_visible_messages"
@@ -213,6 +220,7 @@ depends_on = [
     "mindroom.constants",
     "mindroom.conversation_resolver",
     "mindroom.hooks.sender",
+    "mindroom.knowledge",
     "mindroom.matrix.cache",
     "mindroom.matrix.client",
     "mindroom.matrix.conversation_cache",
@@ -306,6 +314,7 @@ depends_on = [
 [[modules]]
 path = "mindroom.knowledge.manager"
 depends_on = [
+    "mindroom.constants",
     "mindroom.knowledge.chunking",
 ]
 
@@ -363,6 +372,7 @@ from = ["mindroom.memory"]
 expose = [
     "ConversationEventCache",
     "EventCacheWriteCoordinator",
+    "THREAD_HISTORY_CACHE_REJECT_REASON_DIAGNOSTIC",
     "THREAD_HISTORY_DEGRADED_DIAGNOSTIC",
     "THREAD_HISTORY_ERROR_DIAGNOSTIC",
     "THREAD_HISTORY_SOURCE_CACHE",
@@ -372,6 +382,7 @@ expose = [
     "ThreadCacheState",
     "ThreadHistoryResult",
     "normalize_nio_event_for_cache",
+    "thread_cache_rejection_reason",
     "thread_cache_state_is_usable",
     "thread_history_result",
 ]
@@ -458,6 +469,7 @@ expose = [
     "add_room_to_space",
     "create_room",
     "edit_message_result",
+    "ensure_thread_tags_power_level",
     "get_joined_rooms",
     "get_room_members",
     "get_room_name",
@@ -465,8 +477,11 @@ expose = [
     "invite_to_room",
     "join_room",
     "leave_room",
+    "login",
     "matrix_client",
+    "matrix_startup_error",
     "replace_visible_message",
+    "restore_login",
     "send_file_message",
     "send_message_result",
 ]
@@ -479,21 +494,39 @@ visibility = ["mindroom.conversation_resolver"]
 exclusive = true
 
 [[interfaces]]
-expose = ["login", "matrix_startup_error", "restore_login"]
-from = ["mindroom.matrix.client"]
-visibility = ["mindroom.matrix.provisioning", "mindroom.matrix.users"]
-exclusive = true
-
-[[interfaces]]
 expose = [
+    "add_room_to_space",
+    "create_room",
     "create_space",
     "ensure_room_directory_visibility",
     "ensure_room_join_rule",
     "ensure_room_name",
     "ensure_thread_tags_power_level",
+    "get_joined_rooms",
+    "get_room_members",
+    "get_room_name",
+    "invite_to_room",
+    "join_room",
+    "leave_room",
 ]
-from = ["mindroom.matrix.client"]
-visibility = ["mindroom.matrix.rooms"]
+from = ["mindroom.matrix.client_room_admin"]
+visibility = ["mindroom.matrix.client", "mindroom.matrix.rooms"]
+exclusive = true
+
+[[interfaces]]
+expose = [
+    "PermanentMatrixStartupError",
+    "login",
+    "matrix_client",
+    "matrix_startup_error",
+    "restore_login",
+]
+from = ["mindroom.matrix.client_session"]
+visibility = [
+    "mindroom.matrix.client",
+    "mindroom.matrix.provisioning",
+    "mindroom.matrix.users",
+]
 exclusive = true
 
 [[interfaces]]
